+++ lino/ui/__init__.py
@@ -988,12 +988,15 @@ class Site(lino.Site):
             % (u,n,v) for n,v,u in self.using(ui)])
 
     def login(self,username):
+        """
+        """
         #~ self.console_user = self.user_model.objects.get(username=username)
         u = self.user_model.objects.get(username=username)
         from lino.core import requests
-        if u.language:
-            from north.dbutils import set_language
-            set_language(u.language)
-        return requests.BaseRequest(renderer=self.ui.text_renderer,
+        #~ if u.language:
+            #~ from north.dbutils import set_language
+            #~ set_language(u.language)
+        return requests.BaseRequest(
+            renderer=self.ui.text_renderer,
             user=u)